refactoring
add check_account.py


add check_account.py


refactoring


refactoring
