+++ scripts/autooffer/mainnet/transfer_all_to_psc.py
@@ -1,0 +1,4 @@
+from dappctrl_rpc import *
+
+token = get_token(default_password)
+print("\tToken: {}".format(token))
